adding linting, commitlint, and loading animations
