+++ client/db.json
@@ -1,54 +1,54 @@
 {
-    "posts": [
-        {
-            "id": 0,
-            "user": {
-                "name": "Peter Diosdado"
-            },
-            "post": {
-                "comment": "This track is so smooth. Love it. This track is so smooth.",
-                "trackLink": "https://open.spotify.com/track/6WgkOAIOPoumdilSCdYFPB?si=6715b589591e4cbc"
-            }
-        },
-        {
-            "id": 1,
-            "user": {
-                "name": "Dorryen Lopez"
-            },
-            "post": {
-                "comment": "This gets my weeb blood FLOWING.",
-                "trackLink": "https://open.spotify.com/track/7ncPjJBktLEOpstY5JOUMK?si=0f6816849bee44b3"
-            }
-        },
-        {
-            "id": 2,
-            "user": {
-                "name": "Dorryen Lopez"
-            },
-            "post": {
-                "comment": "This gets my weeb blood FLOWING.",
-                "trackLink": "https://open.spotify.com/track/7ncPjJBktLEOpstY5JOUMK?si=0f6816849bee44b3"
-            }
-        },
-        {
-            "id": 3,
-            "user": {
-                "name": "Sam De Leon"
-            },
-            "post": {
-                "comment": "Old head hours.",
-                "trackLink": "https://open.spotify.com/track/4k2UAgNkZ73g3Gsap6TE29?si=6fa5e6d200334f54"
-            }
-        },
-        {
-            "id": 4,
-            "user": {
-                "name": "Kevin Lopez"
-            },
-            "post": {
-                "comment": "This song helped me through tough times.",
-                "trackLink": "https://open.spotify.com/track/3mScGCzxiXA9OaHdBeuk7O?si=d4a5b820388f4b0a"
-            }
-        }
-    ]
+  "posts": [
+    {
+      "id": 0,
+      "user": {
+        "name": "Peter Diosdado"
+      },
+      "post": {
+        "comment": "This track is so smooth. Love it. This track is so smooth.",
+        "trackLink": "https://open.spotify.com/track/6WgkOAIOPoumdilSCdYFPB?si=6715b589591e4cbc"
+      }
+    },
+    {
+      "id": 1,
+      "user": {
+        "name": "Dorryen Lopez"
+      },
+      "post": {
+        "comment": "This gets my weeb blood FLOWING.",
+        "trackLink": "https://open.spotify.com/track/7ncPjJBktLEOpstY5JOUMK?si=0f6816849bee44b3"
+      }
+    },
+    {
+      "id": 2,
+      "user": {
+        "name": "Dorryen Lopez"
+      },
+      "post": {
+        "comment": "This gets my weeb blood FLOWING.",
+        "trackLink": "https://open.spotify.com/track/7ncPjJBktLEOpstY5JOUMK?si=0f6816849bee44b3"
+      }
+    },
+    {
+      "id": 3,
+      "user": {
+        "name": "Sam De Leon"
+      },
+      "post": {
+        "comment": "Old head hours.",
+        "trackLink": "https://open.spotify.com/track/4k2UAgNkZ73g3Gsap6TE29?si=6fa5e6d200334f54"
+      }
+    },
+    {
+      "id": 4,
+      "user": {
+        "name": "Kevin Lopez"
+      },
+      "post": {
+        "comment": "This song helped me through tough times.",
+        "trackLink": "https://open.spotify.com/track/3mScGCzxiXA9OaHdBeuk7O?si=d4a5b820388f4b0a"
+      }
+    }
+  ]
 }
