+++ client/src/App.js
@@ -1,13 +1,14 @@
+/* eslint-disable react/react-in-jsx-scope */
 import './App.css';
-import Header from './components/Header'
+import Header from './components/Header';
 import FeedPage from './pages/FeedPage';
 
 function App() {
   return (
-    <div>
+    <>
       <Header />
       <FeedPage />
-    </div>
+    </>
   );
 }
