--- client/src/pages/LandingPage.js
@@ -1,15 +1,0 @@
-import { Box } from '@mui/system'
-import React from 'react'
-
-export default function LandingPage() {
-    return(
-        <Box sx={{
-            height: '100vh',
-            width: '100vw',
-            backgroundColor: '#414536'
-            }}
-        >
-
-        </Box>
-    )
-}
