+++ client/src/pages/LandingPage.jsx
@@ -1,0 +1,14 @@
+import { Box } from '@mui/system';
+import React from 'react';
+
+export default function LandingPage() {
+  return (
+    <Box
+      sx={{
+        height: '100vh',
+        width: '100vw',
+        backgroundColor: '#414536',
+      }}
+    />
+  );
+}
