+++ client/src/redux/slices/postsState.js
@@ -1,32 +1,33 @@
-import { createSlice,  createAsyncThunk } from "@reduxjs/toolkit";
-import axios from 'axios'
+import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import axios from 'axios';
 
-export const getPosts = createAsyncThunk(
-    'posts/getPosts',
-    async() => {
-        const response = await axios.get('http://localhost:4000/posts')
-        return response.data
-    }
-)
+export const getPosts = createAsyncThunk('posts/getPosts', async () => {
+  const response = await axios.get('http://localhost:4000/posts');
+  return response.data;
+});
 
 export const postsSlice = createSlice({
-    name: 'posts',
-    initialState: {
-        posts: [],
-        isLoading: false
-    },
-    extraReducers: {
-        [getPosts.pending]: (state) => {
-            state.isLoading = true;
-        },
-        [getPosts.fulfilled]: (state, action) => {
-            state.posts = action.payload
-            state.isLoading = false
-        },
-        [getPosts.rejected] : (state) => {
-            state.isLoading = false
-        }
-    }
-})
+  name: 'posts',
+  initialState: {
+    posts: [],
+    isLoading: false,
+  },
 
-export default postsSlice.reducer
+  extraReducers: {
+    [getPosts.pending]: (state) => ({
+      ...state,
+      isLoading: true,
+    }),
+    [getPosts.fulfilled]: (state, { payload }) => ({
+      ...state,
+      posts: payload,
+      isLoading: false,
+    }),
+    [getPosts.rejected]: (state) => ({
+      ...state,
+      isLoading: false,
+    }),
+  },
+});
+
+export default postsSlice.reducer;
